rpm: use system libpmem on Centos 9 Stream
We need libpmem 1.10 and Centos 9 Stream has it.  On top of sticking
to distro-provided packages being generally a good thing, this fixes
a build failure: libpmem 1.10 doesn't build with LTO which is enabled
by default in Centos 9 Stream.  The distro package works around it.

libpmem 1.10 is also there in Fedora 34 and Fedora 35.

Fixes: https://tracker.ceph.com/issues/54473
Signed-off-by: Ilya Dryomov <idryomov@gmail.com>